Modified Server (help)
Modified Server such that new schedules will be sent 1 minute after midnight instead of at exactly minute to prevent any issues with clock desynchronization between Server and Client.
If a Client receives a schedule even a couple of seconds before midnight (according to it's own clock), it would most likely ruin the schedule for the day.

Suggestions for a better fix would be much appreciated.